sdk/python/pyproject.toml
# Python SDK - Client library for capturing IntentEvents

[project]
name = "fencio"
version = "1.2.6"
description = "Fencio SDK - Security enforcement for LangGraph agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "PyYAML>=6.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
]

[project.optional-dependencies]
langgraph = [
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tupl", "fencio"]
artifacts = [
    "tupl/generated/*.py"
]

[tool.hatch.build.targets.sdist]
packages = ["tupl", "fencio"]
artifacts = [
    "tupl/generated/*.py"
]

[tool.hatch.build.targets.wheel.force-include]
"tupl/vocabulary.yaml" = "tupl/vocabulary.yaml"

[tool.hatch.build.targets.sdist.force-include]
"tupl/vocabulary.yaml" = "tupl/vocabulary.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
